+++ package.json
@@ -37,6 +37,9 @@
     "@sentry/react": "^8.17.0",
     "@xstate/react": "^4.1.1",
     "audiomotion-analyzer": "^4.5.0",
+    "change-case": "^5.4.4",
+    "extendable-media-recorder": "^9.2.11",
+    "extendable-media-recorder-wav-encoder": "^7.0.115",
     "xstate": "^5.17.4"
   }
 }
